Source code from the Heroes&Heretics project
This includes an AI script which simulates a battle against another player. This script is currently still in development and contains some temporary code. It is mainly designed around an interruption system which allows the AI to make the best possible decision based on ingame parameters.  

The AudioData script uses the Unity3D FMOD plugin and derives from the scriptable object class to allow designers and programmers to easily edit this as a database script for the audio.

The Audio script handles all of the audio that is used and/or played in the game. This includes support for 3D/2D, announcements, one shot audio and more.